+++ Assets/Cyou/Script/SO Data/Implemented Cl/RedMoonReligionPopeDataSO.cs
@@ -19,11 +19,11 @@ namespace GameData
             populationCost = 5;
 
             // 移動AP
-            moveAPCost = 10f;
+            moveAPCost = 10;
 
             // HP・AP（初始, 升級1, 升級2, 升級3）
-            maxHPByLevel = new float[4] { 3f, 5f, 7f, 7f };
-            maxAPByLevel = new float[4] { 5f, 3f, 3f, 3f };
+            maxHPByLevel = new int[4] { 3, 5, 7, 7 };
+            maxAPByLevel = new int[4] { 5, 3, 3, 3 };
 
             // 各項目のアップグレードコスト (Excel行51-53)
             hpUpgradeCost = new int[3] { 6, 8, 0 }; // 血量: 3→5(6資源), 5→7(8資源), 升級3なし
